+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added support for array arithmetic. (#628)
+- Added ARRAYFORMULA function. (#630)
 - Added performance improvements for array handling. (#629)
 - Accepting time in JS Date() objects on the input. (#648)